Move useIncrementalResponse hook to libs/ajax
