+++ src/libs/ajax/incremental-response/IncrementalResponse.ts
