+++ src/libs/ajax/incremental-response/useIncrementalResponse.test.ts
@@ -1,8 +1,9 @@
 import { act, renderHook } from '@testing-library/react-hooks'
 import _ from 'lodash/fp'
-import useIncrementalResponse from 'src/components/common/hooks/useIncrementalResponse'
-import IncrementalResponse from 'src/libs/ajax/IncrementalResponse'
 import { ErrorState, LoadingState, ReadyState } from 'src/libs/type-utils/LoadedState'
+
+import IncrementalResponse from './IncrementalResponse'
+import useIncrementalResponse from './useIncrementalResponse'
 
 
 describe('useIncrementalResponse', () => {
